+++ CHANGES.md
@@ -1,3 +1,7 @@
+### V 1.1.0 29/04/2016
+ * SMServer service is now started at the end of Android boot phase
+ * Add SMServer version in notification bar
+
 ### V 1.0.0 06/09/2015
  * First release